Update of the Server RHEL 9.4 and fixing related issues
    RHEL 9.4
    gcc 11
    CUDA 11.8

Possible configuration of CUDA and alternative of using 12 in cmake

CUDA 11.8 is the lowest possible version of CUDA supporting platform
Kepler, compute 35.

For CUDA 12.1 there are additional architectures, for CUDA 11.8 there is
35 enabled. It could be selected by changing

set(CUDA_VERSION "11.8")
or
set(CUDA_VERSION "12.1")

if(CUDAToolkit_VERSION VERSION_LESS "12.0")
    # CUDA 11.x architecture support including Kepler 35 and Maxwell 50 52 Kepler
    # is K40 still on Maxwell
    set(CUDA_ARCHITECTURES_LIST "35;50;52;60;61;70;75")
else()
    # CUDA 12.x architecture support
    set(CUDA_ARCHITECTURES_LIST "50;52;60;61;70;75;80;86")
endif()

For set(CUDA_VERSION "12.1")

module load cuda/12.1
module load gcc/12.2

for  set(CUDA_VERSION "11.8")

module load cuda/11.8

+++ src/dentk-basis.cpp
@@ -51,13 +51,13 @@ void shiftBasis(std::string inputFile, std::string outputFile, int shift)
     for(uint32_t i = 0; i != baseSize; i++)
     {
         f = denSliceReader->readFrame(i);
-        for(int j = 0; j != granularity; j++)
+        for(uint32_t j = 0; j != granularity; j++)
         {
             int index = j + shift;
             if(index < 0)
             {
                 index = 0;
-            } else if(index >= granularity)
+            } else if(static_cast<uint32_t>(index) >= granularity)
             {
                 index = granularity - 1;
             }
@@ -86,7 +86,7 @@ void resampleFunctions(std::string inputFile, std::string outputFile, uint16_t r
     double* newValues = new double[resampledGranularity];
     MKL_INT bc_type = DF_BC_1ST_LEFT_DER | DF_BC_1ST_RIGHT_DER;
     double bc[2] = { 0.0, 0.0 };
-    for(int i = 0; i != baseSize; i++)
+    for(uint32_t i = 0; i != baseSize; i++)
     {
         f = denSliceReader->readFrame(i);
         for(int j = 0; j != granularity; j++)
